Exported auto complete to own file
